[#9] refactor: enums, consts 패키지 위치 수정
