+++ core/src/main/kotlin/site/snaplog/util/consts/ColorCode.kt
@@ -1,4 +1,4 @@
-package site.snaplog.util
+package site.snaplog.util.consts
 
 object ColorCode {
