+++ core/src/main/kotlin/site/snaplog/util/enums/Provider.kt
@@ -1,4 +1,4 @@
-package site.snaplog.enums
+package site.snaplog.util.enums
 
 enum class Provider {
     GOOGLE,
